requirements.txt
python-dotenv
datasets
sentence-transformers
chromadb
google-genai
streamlit
uvicorn
fastapi
pydantic
transformers
tqdm
